refactor: TraceSubscriptionService to TraceSubscriptionManager
